fix(contacts): pass usize to DataTable.total_items
DataTable.total_items is typed usize but CompanyListPage cast remote_companies.len() to u32, breaking the wasm release build (E0308 mismatched types). slice::len() returns usize so the cast is also unneeded - drop it.

+++ src/pages/contacts.rs
@@ -135,7 +135,7 @@ pub fn CompanyListPage() -> Element {
 
             // Companies table
             DataTable {
-                total_items: if source == CompanySource::Backend { remote_companies.len() as u32 } else { 5 },
+                total_items: if source == CompanySource::Backend { remote_companies.len() } else { 5 },
                 current_page: 1,
                 per_page: 25,
                 columns: 5,
